+++ CHANGES.rst
@@ -6,7 +6,11 @@ Changelog
 1.2.71 (unreleased)
 -------------------
 
-- Nothing changed yet.
+Added:
+
+- The help bubbles behavior can now be controlled through the registry record
+  ``ploneintranet.layout.bubbles_enabled`` (Refs. Syslab #15824).
+  [Alessandrop Pisa]
 
 
 1.2.70 (2017-08-28)